Parking-App/requirements.txt
dotenv~=0.9.9
python-dotenv~=1.1.1
Flask==2.3.2
mysql-connector-python==9.2.0
python-dotenv==1.0.0
gunicorn==23.0.0